requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39
pyyaml>=6.0
pytest>=7.0.0